Remove "multiplayer" references from subclass and move to appropriate place
